COMP: Eliminate some unused variable and signed/unsigned comparison       warnings.

+++ Utilities/vtkexodus2/exgsnl.c
@@ -288,7 +288,7 @@ int ex_get_side_set_node_list_len(int exoid,
     elem_blk_parms[i].num_attr = num_attr;
     elem_blk_parms[i].elem_blk_id = elem_blk_ids[i];
 
-    for (m=0; m < strlen(elem_type); m++)
+    for (m=0; m < (int)strlen(elem_type); m++)
       elem_blk_parms[i].elem_type[m] = 
               toupper((int)elem_type[m]);
     elem_blk_parms[i].elem_type[m] = '\0';
