+++ Utilities/vtkexodus2/exgssn.c
@@ -378,7 +378,7 @@ int ex_get_side_set_node_list(int exoid,
     elem_blk_parms[i].num_nodes_per_elem = num_nodes_per_elem;
     elem_blk_parms[i].num_attr = num_attr;
 
-    for (m=0; m < strlen(elem_type); m++)
+    for (m=0; m < (int)strlen(elem_type); m++)
       elem_blk_parms[i].elem_type[m] = 
               toupper((int)elem_type[m]);
     elem_blk_parms[i].elem_type[m] = '\0';
